+++ lib/react_on_rails/helper.rb
@@ -158,7 +158,17 @@ module ReactOnRails
     #    immediate_hydration: nil -- React on Rails Pro (licensed) feature. When nil (default), Pro users
     #                        get immediate hydration, non-Pro users don't. Can be explicitly overridden.
     def redux_store(store_name, props: {}, defer: false, immediate_hydration: nil)
-      immediate_hydration = ReactOnRails::Utils.react_on_rails_pro? if immediate_hydration.nil?
+      # If non-Pro user explicitly sets immediate_hydration: true, warn and override to false
+      if immediate_hydration == true && !ReactOnRails::Utils.react_on_rails_pro?
+        Rails.logger.warn <<~WARNING
+          [REACT ON RAILS] Warning: immediate_hydration: true requires a React on Rails Pro license.
+          Store '#{store_name}' will fall back to standard hydration behavior.
+          Visit https://www.shakacode.com/react-on-rails-pro/ for licensing information.
+        WARNING
+        immediate_hydration = false
+      elsif immediate_hydration.nil?
+        immediate_hydration = ReactOnRails::Utils.react_on_rails_pro?
+      end
 
       redux_store_data = { store_name: store_name,
                            props: props,